+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Cover more functional tests on back end Groups page
 - Add button on back end add forms are now disabled while saving
 - Expose instances of Models through the Core API. It is now possible to manipulate taxonomies, Web Service clients, groups, roles and users through the Core API
+- Add new Angular Formly types "ovFile" and "horizontalFile" to add files. Actually it gives you the possibility to add a file with a progress bar (without preview)
 
 ## DEPENDENCIES